Adding support for booting Windows on Arm.
Note however that this isn't the preferable way for VHE, as with VHE,
the kernel uses the EL2 physical timers, which would allow us to
directly expose the EL1 physical timer to the guest.

On Carmel, the resulting Windows on Arm VM is very slow and slows
down the host too. A cache flushing or TLB issue is suspected.
Shutting the VM down makes everything return as normal. This
does not affect Linux VMs.

On the documentation side: removing the signing instructions
as they are not required with CBoot. Note also that installing
the qemu-user(-static) package is enough to make those tools
directly run on a Xavier.